const.tsとtype.tsファイルを削除し、関連するインポートをconstantsとtypesに変更しました。また、PartCreateMenu、PartPropertySidebar、GameBoardコンポーネントでのインポートを更新しました。
