+++ frontend/src/features/prototype/components/molecules/PartCreateMenu.tsx
@@ -13,9 +13,9 @@ import {
 } from 'react-icons/gi';
 
 import { Part, PartProperty } from '@/api/types';
-import { PART_DEFAULT_CONFIG } from '@/features/prototype/const';
+import { PART_DEFAULT_CONFIG } from '@/features/prototype/constants';
 import { CANVAS_SIZE } from '@/features/prototype/constants';
-import { AddPartProps } from '@/features/prototype/type';
+import { AddPartProps } from '@/features/prototype/types';
 import { isRectOverlap } from '@/features/prototype/utils/overlap';
 
 export default function PartCreateMenu({
