+++ frontend/src/features/prototype/components/organisms/GameBoard.tsx
@@ -30,7 +30,7 @@ import { useHandVisibility } from '@/features/prototype/hooks/useHandVisibility'
 import { usePartReducer } from '@/features/prototype/hooks/usePartReducer';
 import { usePerformanceTracker } from '@/features/prototype/hooks/usePerformanceTracker';
 import { useSelection } from '@/features/prototype/hooks/useSelection';
-import { AddPartProps, DeleteImageProps } from '@/features/prototype/type';
+import { AddPartProps, DeleteImageProps } from '@/features/prototype/types';
 import { CursorInfo } from '@/features/prototype/types';
 import { GameBoardMode } from '@/features/prototype/types';
 import { useRoleManagement } from '@/features/role/hooks/useRoleManagement';
