+++ frontend/src/features/prototype/constants/index.ts
@@ -1,2 +1,3 @@
 export * from './gameBoard';
 export * from './helpInfo';
+export * from './part';
