+++ frontend/src/features/prototype/constants/part.ts
@@ -1,11 +1,30 @@
-import { PartDefaultConfig } from './type';
+import { PartDefaultConfig } from '../type';
 
-// グリッドサイズ
-export const GRID_SIZE = 1;
-
-// プレイヤー数の最小値と最大値
-export const PLAYERS_MIN = 1;
-export const PLAYERS_MAX = 12;
+/**
+ * パーツのカラー定義
+ */
+export const COLORS = {
+  /** 背景色 */
+  BACKGROUNDS: [
+    '#FF0000', // Red
+    '#00FF00', // Green
+    '#0000FF', // Blue
+    '#FFFF00', // Yellow
+    '#FF00FF', // Magenta
+    '#00FFFF', // Cyan
+    '#FFA500', // Orange
+    '#800080', // Purple
+    '#008000', // Dark Green
+    '#FFC0CB', // Pink
+    '#FFFFFF', // White
+    '#808080', // Gray
+  ],
+  /** テキスト色 */
+  TEXT: [
+    '#FFFFFF', // White
+    '#000000', // Black
+  ],
+};
 
 // パーツのデフォルト設定
 export const PART_DEFAULT_CONFIG: Record<string, PartDefaultConfig> = {
@@ -62,32 +81,3 @@ export const PART_DEFAULT_CONFIG: Record<string, PartDefaultConfig> = {
     textColor: '#000000',
   },
 };
-
-// カラー
-export const COLORS = [
-  '#FF0000', // Red
-  '#00FF00', // Green
-  '#0000FF', // Blue
-  '#FFFF00', // Yellow
-  '#FF00FF', // Magenta
-  '#00FFFF', // Cyan
-  '#FFA500', // Orange
-  '#800080', // Purple
-  '#008000', // Dark Green
-  '#FFC0CB', // Pink
-  '#FFFFFF', // White
-  '#808080', // Gray
-];
-
-// テキスト色
-export const TEXT_COLORS = [
-  '#FFFFFF', // White
-  '#000000', // Black
-];
-
-// プレイヤー名の色
-export const PLAYER_NAME_COLORS = [
-  '#FFFFFF', // White
-  '#808080', // Gray
-  '#000000', // Black
-];
