+++ frontend/src/features/prototype/types/index.ts
@@ -1,3 +1,4 @@
 export * from './gameBoardMode';
 export * from './cursor';
 export * from './socket';
+export * from './part';
